rm old circuits, organize new circuits
