+++ src/compiled/lay_egg.rs
