+++ src/compiled/pickup_egg.rs
