--- src_c/karma_add/karma_add.h
@@ -1,6 +1,0 @@
-#ifndef FULLY_HOMOMORPHIC_ENCRYPTION_TRANSPILER_EXAMPLES_FIBONACCI_FIBONACCI_H_
-#define FULLY_HOMOMORPHIC_ENCRYPTION_TRANSPILER_EXAMPLES_FIBONACCI_FIBONACCI_H_
-
-unsigned short karma_add(unsigned short a, unsigned short b);
-
-#endif  // FULLY_HOMOMORPHIC_ENCRYPTION_TRANSPILER_EXAMPLES_FIBONACCI_FIBONACCI_H_
