--- src_c/karma_add/karma_add_rs_lib.rs
@@ -1,4 +1,0 @@
-use karma_add_rs_fhe_lib::karma_add;
-
-#[cfg(not(test))]
-fn main() {}
