--- src_c/karma_sub/karma_sub.h
@@ -1,6 +1,0 @@
-#ifndef FULLY_HOMOMORPHIC_ENCRYPTION_TRANSPILER_EXAMPLES_FIBONACCI_FIBONACCI_H_
-#define FULLY_HOMOMORPHIC_ENCRYPTION_TRANSPILER_EXAMPLES_FIBONACCI_FIBONACCI_H_
-
-unsigned short karma_sub(unsigned short a, unsigned short b);
-
-#endif  // FULLY_HOMOMORPHIC_ENCRYPTION_TRANSPILER_EXAMPLES_FIBONACCI_FIBONACCI_H_
