--- src_c/karma_sub/karma_sub_rs_lib.rs
@@ -1,4 +1,0 @@
-use karma_sub_rs_fhe_lib::karma_sub;
-
-#[cfg(not(test))]
-fn main() {}
